factor out L2 projection code into new class FEProjector
