+++ ibtk/lib/Makefile.am
@@ -190,6 +190,7 @@ DIM_DEPENDENT_SOURCES += \
 ../src/lagrangian/JacobianCalculator.cpp \
 ../src/lagrangian/FEDataInterpolation.cpp \
 ../src/lagrangian/FEDataManager.cpp \
+../src/lagrangian/FEProjector.cpp \
 ../src/utilities/LibMeshSystemIBVectors.cpp \
 ../src/utilities/LibMeshSystemVectors.cpp \
 ../src/utilities/libmesh_utilities.cpp
@@ -355,12 +356,15 @@ pkg_include_HEADERS += \
 ../include/ibtk/private/PETScSAMRAIVectorReal-inl.h \
 ../include/ibtk/private/StreamableManager-inl.h
 
+## TODO: Should JacobianCalculator.h be in include or src?
+
 if LIBMESH_ENABLED
 DIM_DEPENDENT_SOURCES += \
 ../include/lagrangian/BoxPartitioner.h \
 ../src/lagrangian/JacobianCalculator.h \
 ../include/ibtk/FEDataInterpolation.h \
 ../include/ibtk/FEDataManager.h \
+../include/ibtk/FEProjector.h \
 ../include/ibtk/LibMeshSystemIBVectors.h \
 ../include/ibtk/LibMeshSystemVectors.h \
 ../include/ibtk/libmesh_utilities.h
